added LRU cache
One bug happened earlier was that - in update frequency, the data doesn't move if there is no data with 1 more frequency. Instead, the logic should be - if there are existing data with frequency same as the updated frequency ( freq=freq+1), this data should be moved to the front of all those; but if there is no existing data with the frequency same as the updated frequency, it still should be moved to the front of all the data that has frequency same as the frequency from updating. See L96